+++ src/SharpX/Primitives.cs
@@ -58,4 +58,23 @@ public static class Primitives
             yield return generator();
         }
     }
+
+    /// <summary>Generates a random sequence of int, double or string types.</summary>
+    public static IEnumerable<T> GenerateSeq<T>(int? count = null)
+    {
+        if (count != null) Guard.DisallowNegative(nameof(count), count.Value);
+
+        Func<object> generator = typeof(T) switch {
+            var t when t == typeof(int) => () => RandomNumberGenerator.GetInt32(int.MinValue, int.MaxValue),
+            var t when t == typeof(double) => () => BitConverter.ToDouble(RandomNumberGenerator.GetBytes(8), 0),
+            var t when t == typeof(string) => () => Strings.Generate(length: 16),
+            _ => throw new ArgumentException($"{typeof(T).Name} is not supported"),
+        };
+
+        var count_ = count ?? 0;
+
+        while (count == null || count_-- > 0) {
+            yield return (T)generator();
+        }
+    }
 }